jarvis_bot/requirements.txt
openai
requests
tk
pillow
pyttsx3
speechrecognition
pyaudio
numpy
python-dotenv
sqlite3
